sqrt(i) --> sqrt((double)i)

+++ OpenMPvalidation/SRC/C/TESTS/for_private.c
@@ -10,7 +10,7 @@ do_some_work ()
   double sum = 0;
   for (i = 0; i < 1000; i++)
     {
-      sum += sqrt (i);
+      sum += sqrt ((double)i);
     }
 }
